+++ relay-server/src/actors/project.rs
@@ -1174,7 +1174,6 @@ impl Project {
 mod tests {
     use std::sync::{Arc, Mutex};
 
-    use itertools::Itertools;
     use relay_common::time::UnixTimestamp;
     use relay_metrics::BucketValue;
     use relay_test::mock_service;
@@ -1433,117 +1432,5 @@ mod tests {
         let metrics = project.rate_limit_metrics(vec![create_transaction_bucket()], addr);
 
         assert!(metrics.is_empty());
-    }
-
-    fn get_test_buckets(names: &[&str]) -> Vec<Bucket> {
-        let create_bucket = |name: &&str| -> Bucket {
-            let json = json!({
-                        "timestamp": 1615889440,
-                        "width": 10,
-                        "name": name,
-                        "type": "c",
-                        "value": 4.0,
-                        "tags": {
-                        "route": "user_index"
-            }});
-
-            serde_json::from_value(json).unwrap()
-        };
-
-        names.iter().map(create_bucket).collect_vec()
-    }
-
-    fn get_test_bucket_names() -> Vec<&'static str> {
-        [
-            "g:transactions/foo@none",
-            "c:custom/foo@none",
-            "transactions/foo@second",
-            "transactions/foo",
-            "c:custom/foo_bar@none",
-            "endpoint.response_time",
-            "endpoint.hits",
-            "endpoint.parallel_requests",
-            "endpoint.users",
-        ]
-        .into()
-    }
-
-    fn apply_pattern_to_names(names: &[&str], patterns: &[&str]) -> Vec<String> {
-        let mut buckets = get_test_buckets(names);
-        let patterns = patterns.iter().map(|s| String::from(*s)).collect_vec();
-        let deny_list = ErrorBoundary::Ok(Metrics::new(patterns));
-        Project::apply_metrics_deny_list(&deny_list, &mut buckets);
-        buckets.into_iter().map(|bucket| bucket.name).collect_vec()
-    }
-
-    #[test]
-    fn test_metric_deny_list_exact() {
-        let names = get_test_bucket_names();
-        let input_qty = names.len();
-        let remaining_names = apply_pattern_to_names(&names, &["endpoint.parallel_requests"]);
-
-        // There's 1 bucket with that exact name.
-        let buckets_to_remove = 1;
-
-        assert_eq!(remaining_names.len(), input_qty - buckets_to_remove);
-    }
-
-    #[test]
-    fn test_metric_deny_list_end_glob() {
-        let names = get_test_bucket_names();
-        let input_qty = names.len();
-        let remaining_names = apply_pattern_to_names(&names, &["*foo"]);
-
-        // There's 1 bucket name with 'foo' in the end.
-        let buckets_to_remove = 1;
-
-        assert_eq!(remaining_names.len(), input_qty - buckets_to_remove);
-    }
-
-    #[test]
-    fn test_metric_deny_list_middle_glob() {
-        let names = get_test_bucket_names();
-        let input_qty = names.len();
-        let remaining_names = apply_pattern_to_names(&names, &["*foo*"]);
-
-        // There's 4 bucket names with 'foo' in the middle, and one with foo in the end.
-        let buckets_to_remove = 5;
-
-        assert_eq!(remaining_names.len(), input_qty - buckets_to_remove);
-    }
-
-    #[test]
-    fn test_metric_deny_list_beginning_glob() {
-        let names = get_test_bucket_names();
-        let input_qty = names.len();
-        let remaining_names = apply_pattern_to_names(&names, &["endpoint*"]);
-
-        // There's 4 buckets starting with "endpoint".
-        let buckets_to_remove = 4;
-
-        assert_eq!(remaining_names.len(), input_qty - buckets_to_remove);
-    }
-
-    #[test]
-    fn test_metric_deny_list_everything() {
-        let names = get_test_bucket_names();
-        let remaining_names = apply_pattern_to_names(&names, &["*"]);
-
-        assert_eq!(remaining_names.len(), 0);
-    }
-
-    #[test]
-    fn test_metric_deny_list_multiple() {
-        let names = get_test_bucket_names();
-        let input_qty = names.len();
-        let remaining_names = apply_pattern_to_names(&names, &["endpoint*", "*transactions*"]);
-
-        let endpoint_buckets = 4;
-        let transaction_buckets = 3;
-
-        assert_eq!(
-            remaining_names.len(),
-            input_qty - endpoint_buckets - transaction_buckets
-        );
     }
 }
